+++ apys/routes.py
@@ -16,6 +16,7 @@ from apys import log, settings
 # CONSTANTS
 #
 
+ATTR_FILTERS = settings.DIR_FILTERS
 ATTR_UTILS = settings.DIR_UTILS
 
 #
@@ -53,6 +54,7 @@ def prepare(app, api, cors_url=False):
     routes = []
     
     utils = {}
+    filters = {}
     
     api.debug('')
     if cors_url:
@@ -60,42 +62,63 @@ def prepare(app, api, cors_url=False):
     else:
         api.debug('================== Resources ====================')
 
-    #populate routes
+    api.debug('')
+    
+    ##
+    # ADDING UTILS TO API PARAM
+    #
+    if os.path.exists(os.path.join('.', settings.DIR_UTILS)):
+        for subdir in os.listdir(os.path.join('.', settings.DIR_UTILS)):
+            if((os.path.isdir(os.path.join('.', settings.DIR_UTILS, subdir))) and
+                (os.path.exists(os.path.join('.', settings.DIR_UTILS, subdir, '__init__.py')))):
+                util = subdir
+                if util not in utils:
+                    utils[util] = imp.load_source(util, os.path.join('.', settings.DIR_UTILS, util, '__init__.py'))
+                    setattr(api, util, utils[util])
+                
+                    # calls util init function 
+                    if hasattr(utils[util], 'init'):
+                        utils[util].init(api)
+
+    # populate routes
     for file_path in file_paths:
-        
-        # api.debug('Loading endpoint: [{}{}{}]'.format(api._bcolors.HEADER, file_path['url'], api._bcolors.ENDC))
-        
+
         file_module = imp.load_source(file_path['url'].replace('/', '-'), file_path['file'])
 
         supported_methods = api.supported_methods
 
         handler_props = {
-            '_params': None,
-            
             'api': api
         }
+
+        loaded_methods = []
+        loaded_filters = {}
         
         ##
-        # ADDING UTILS TO API PARAM
+        # ADDING FILTER TO LIST
         #
-        if hasattr(file_module, ATTR_UTILS):
-            for util in file_module.utils:
-                if not util in utils:
-                    utils[util] = imp.load_source(util, os.path.join(settings.DIR_UTILS, util, '__init__.py'))
-                setattr(handler_props['api'], util, utils[util])
-                
-                #calls util init function 
-                if hasattr(utils[util], 'init'):
-                    utils[util].init(api)
+        if hasattr(file_module, ATTR_FILTERS):
+            for filt in file_module.filters:
+                if not filt in filters:
+                    filters[filt] = imp.load_source(filt, '{}.py'.format(os.path.join(settings.DIR_FILTERS, *(filt.split('.')))))
+
+                    #calls util init function 
+                    if hasattr(filters[filt], 'init'):
+                        filters[filt].init(api)
+
+                loaded_filters[filt] = filters[filt]
 
         ##
         # ADDING METHOD FUNCTIONS
         #
         def createFunc(endpoint, method):
+
+            ##
+            # This is what gets executed when you call the endpoint
+            #
             async def func(req):
                 req.vars = {}
                 try:
-                    api.debug(req)
                     if req.has_body:
                         req.body = await req.json()
                         req.body = __translateJson(req.body)
@@ -104,14 +127,14 @@ def prepare(app, api, cors_url=False):
                 except Exception as ex:
                     api.error('Error while getting json data', ex=ex)
                 
-                if hasattr(endpoint, ATTR_UTILS):
-                    for util in endpoint.utils:
-                        # Calls 'any' util function
-                        if hasattr(utils[util], 'any'):
-                            utils[util].any(req, handler_props['api'])
-                        # Calls current method util function
-                        if hasattr(utils[util], method):
-                            getattr(utils[util], method)(req, handler_props['api'])
+                if hasattr(endpoint, ATTR_FILTERS):
+                    for filt in endpoint.filters:
+                        # Calls 'any' filt function
+                        if hasattr(filters[filt], 'any'):
+                            filters[filt].any(req, handler_props['api'])
+                        # Calls current method filt function
+                        if hasattr(filters[filt], method):
+                            getattr(filters[filt], method)(req, handler_props['api'])
 
                 # Calls current method function
                 
@@ -129,7 +152,6 @@ def prepare(app, api, cors_url=False):
             resource = cors.add(app.router.add_resource('/{}'.format(file_path['url'])))
         else: 
             resource = app.router.add_resource('/{}'.format(file_path['url']))
-        loaded_methods = []
         for method in supported_methods:
             if hasattr(file_module, method):
 
@@ -141,24 +163,26 @@ def prepare(app, api, cors_url=False):
                     resource.add_route(method.upper(), createFunc(file_module, method))
         
         # Logging loaded endpoints
-        str_loaded_methods = '('
         for method in loaded_methods:
-            str_loaded_methods += api._bcolors.OKGREEN + method + api._bcolors.ENDC
-            str_loaded_methods += ', '
-        str_loaded_methods = str_loaded_methods[0:-2]
-        str_loaded_methods += ')'
-        api.debug('Endpoint Loaded: [{}{}{}] {}'.format(api._bcolors.OKGREEN, file_path['url'], api._bcolors.ENDC, str_loaded_methods))
+            str_loaded_methods = '{}{}{}'.format(api._bcolors.OKGREEN, method, api._bcolors.ENDC)
+
+            str_loaded_filters = ''
+            for filt in loaded_filters:
+                if hasattr(loaded_filters[filt], 'any'):
+                    str_loaded_filters += '{}{}{}'.format(api._bcolors.OKBLUE, filt, api._bcolors.ENDC)
+                    str_loaded_filters += ', '
+                if hasattr(loaded_filters[filt], method):
+                    str_loaded_filters += '{}{}{}'.format(api._bcolors.OKGREEN, filt, api._bcolors.ENDC)
+                    str_loaded_filters += ', '
+            str_loaded_filters = str_loaded_filters[0:-2]
+            
+            api.debug('Endpoint Loaded: [{}{}{}] ({}) <- {{{}}}'.format(api._bcolors.OKGREEN, file_path['url'], api._bcolors.ENDC, str_loaded_methods, str_loaded_filters))
     
+    api.debug('')
+
     # Logging loades utils
-    if os.path.exists(os.path.join('.', settings.DIR_UTILS)):
-        for subdir in os.listdir(os.path.join('.', settings.DIR_UTILS)):
-            if((os.path.isdir(os.path.join('.', settings.DIR_UTILS, subdir))) and
-                (os.path.exists(os.path.join('.', settings.DIR_UTILS, subdir, '__init__.py')))):
-                util = subdir
-                if util in utils:
-                    api.debug('Util Loaded: [{}{}{}]'.format(api._bcolors.OKBLUE, util, api._bcolors.ENDC))
-                else:
-                    api.debug('Util not Loaded: [{}{}{}]'.format(api._bcolors.WARNING, util, api._bcolors.ENDC))
+    for util in utils:
+        api.debug('Util Loaded: [{}{}{}]'.format(api._bcolors.OKBLUE, util, api._bcolors.ENDC))
     
     api.debug('')
     return app